requirements.txt
keras==2.9.0
keras_nightly==2.5.0.dev2021032900
mediapipe==0.8.10.1
numpy==1.22.2
Pillow==9.2.0
streamlit==1.8.1
opencv_python==4.5.2.52
